moving ide_{get,update}_ftd_var here
they were initially in `dotcom` repo. we need them here because we
are going to now use the preact thing to power the package list,
and we want to do it in such a way that there is no modification needed
in dotcom repo.

we will intercept the calls to set package data variable and call
preact code.