Wait for 1s before saving user input to opfs
This wait time is not final and can change in future. I personally find
that 2s is just too slow to see UI react to my changes (save button
appearing, preview becoming stale etc) so I've decided to change it to
1s.

This value should be as small as possible but we also want to not burn
user's CPU and battery so 1s might be fine.